moves v2 repositories to legacy->two folder
