+++ app/Http/Controllers/Legacy/Two/TagsController.php
@@ -2,7 +2,7 @@
 
 namespace Rogue\Http\Controllers\Legacy\Two;
 
-use Rogue\Repositories\PostRepository;
+use Rogue\Repositories\Legacy\Two\PostRepository;
 use Rogue\Http\Transformers\Legacy\Two\PostTransformer;
 use Rogue\Http\Controllers\Traits\TagsRequests;
 
@@ -12,7 +12,7 @@ class TagsController extends ApiController
     /**
      * The post service instance.
      *
-     * @var Rogue\Repositories\PostRepository
+     * @var Rogue\Repositories\Legacy\Two\PostRepository
      */
     protected $post;
